Step1 피드백 반영 | Car, InputView, ResultView 클래스 분리
