+++ src/main/java/racing/Car.java
@@ -2,10 +2,10 @@ package racing;
 
 public class Car {
 
+    private static final int CRITERION_TO_MOVE = 4;
     private int position = 0;
 
     public Car() {}
-
     public Car(int position){
         this.position = position;
     }
@@ -13,9 +13,14 @@ public class Car {
     public int getPosition() {
         return position;
     }
-
     public void setPosition(int position) {
         this.position = position;
+    }
+
+    public void move(int randomValue) {
+        if (randomValue > CRITERION_TO_MOVE) {
+            this.position++;
+        }
     }
 
 }
