+++ src/main/java/racing/RacingGameMain.java
@@ -1,10 +1,21 @@
 package racing;
 
+import java.util.List;
+
 public class RacingGameMain {
     public static void main(String args[]) {
 
-        RacingGame.inputView();
-        RacingGame.initCars();
-        RacingGame.resultView();
+        int numberOfCars = InputView.getNumberOfCar();
+        int timeToTry = InputView.getTimeToTry();
+
+        List<Car> cars = RacingGame.initCars(numberOfCars);
+
+        int time = 0;
+        while(time < timeToTry) {
+            cars = RacingGame.race(cars);
+            ResultView.printResult(cars);
+            time++;
+        }
+
     }
 }
